refactor: remove encode/decode for command block types
This logic was superfluous. We now hand raw types to the state layer,
and do not offset from the iota to match database lookup keys.

No validation is required for incoming type IDs, because inserts will
fail for unknown types.

+++ domain/blockcommand/state/state_test.go
@@ -23,38 +23,38 @@ func TestStateSuite(t *testing.T) {
 
 func (s *stateSuite) TestSetBlock(c *tc.C) {
 	st := NewState(s.TxnRunnerFactory())
-	err := st.SetBlock(c.Context(), blockcommand.DestroyBlock, "block-message")
+	err := st.SetBlock(c.Context(), int8(blockcommand.DestroyBlock), "block-message")
 
 	c.Assert(err, tc.ErrorIsNil)
 }
 
 func (s *stateSuite) TestSetBlockForSameTypeTwice(c *tc.C) {
 	st := NewState(s.TxnRunnerFactory())
-	err := st.SetBlock(c.Context(), blockcommand.DestroyBlock, "block-message")
+	err := st.SetBlock(c.Context(), int8(blockcommand.DestroyBlock), "block-message")
 	c.Assert(err, tc.ErrorIsNil)
-	err = st.SetBlock(c.Context(), blockcommand.DestroyBlock, "block-message")
+	err = st.SetBlock(c.Context(), int8(blockcommand.DestroyBlock), "block-message")
 	c.Assert(err, tc.ErrorIs, blockcommanderrors.AlreadyExists)
 }
 
 func (s *stateSuite) TestSetBlockWithNoMessage(c *tc.C) {
 	st := NewState(s.TxnRunnerFactory())
-	err := st.SetBlock(c.Context(), blockcommand.DestroyBlock, "")
+	err := st.SetBlock(c.Context(), int8(blockcommand.DestroyBlock), "")
 	c.Assert(err, tc.ErrorIsNil)
 }
 
 func (s *stateSuite) TestRemoveBlockWithNoExistingBlock(c *tc.C) {
 	st := NewState(s.TxnRunnerFactory())
-	err := st.RemoveBlock(c.Context(), blockcommand.DestroyBlock)
+	err := st.RemoveBlock(c.Context(), int8(blockcommand.DestroyBlock))
 
 	c.Assert(err, tc.ErrorIs, blockcommanderrors.NotFound)
 }
 
 func (s *stateSuite) TestRemoveBlockWithExistingBlock(c *tc.C) {
 	st := NewState(s.TxnRunnerFactory())
-	err := st.SetBlock(c.Context(), blockcommand.DestroyBlock, "")
+	err := st.SetBlock(c.Context(), int8(blockcommand.DestroyBlock), "")
 	c.Assert(err, tc.ErrorIsNil)
 
-	err = st.RemoveBlock(c.Context(), blockcommand.DestroyBlock)
+	err = st.RemoveBlock(c.Context(), int8(blockcommand.DestroyBlock))
 	c.Assert(err, tc.ErrorIsNil)
 }
 
@@ -69,9 +69,9 @@ func (s *stateSuite) TestGetBlocksWithNoExistingBlock(c *tc.C) {
 func (s *stateSuite) TestGetBlocks(c *tc.C) {
 	st := NewState(s.TxnRunnerFactory())
 
-	err := st.SetBlock(c.Context(), blockcommand.DestroyBlock, "")
+	err := st.SetBlock(c.Context(), int8(blockcommand.DestroyBlock), "")
 	c.Assert(err, tc.ErrorIsNil)
-	err = st.SetBlock(c.Context(), blockcommand.ChangeBlock, "change me")
+	err = st.SetBlock(c.Context(), int8(blockcommand.ChangeBlock), "change me")
 	c.Assert(err, tc.ErrorIsNil)
 
 	blocks, err := st.GetBlocks(c.Context())
@@ -86,7 +86,7 @@ func (s *stateSuite) TestGetBlocks(c *tc.C) {
 
 func (s *stateSuite) TestGetBlockMessageWithNoExistingBlock(c *tc.C) {
 	st := NewState(s.TxnRunnerFactory())
-	message, err := st.GetBlockMessage(c.Context(), blockcommand.DestroyBlock)
+	message, err := st.GetBlockMessage(c.Context(), int8(blockcommand.DestroyBlock))
 
 	c.Assert(err, tc.ErrorIs, blockcommanderrors.NotFound)
 	c.Assert(message, tc.Equals, "")
@@ -94,10 +94,10 @@ func (s *stateSuite) TestGetBlockMessageWithNoExistingBlock(c *tc.C) {
 
 func (s *stateSuite) TestGetBlockMessage(c *tc.C) {
 	st := NewState(s.TxnRunnerFactory())
-	err := st.SetBlock(c.Context(), blockcommand.DestroyBlock, "destroy me")
+	err := st.SetBlock(c.Context(), int8(blockcommand.DestroyBlock), "destroy me")
 	c.Assert(err, tc.ErrorIsNil)
 
-	message, err := st.GetBlockMessage(c.Context(), blockcommand.DestroyBlock)
+	message, err := st.GetBlockMessage(c.Context(), int8(blockcommand.DestroyBlock))
 
 	c.Assert(err, tc.ErrorIsNil)
 	c.Assert(message, tc.Equals, "destroy me")
@@ -112,7 +112,7 @@ func (s *stateSuite) TestRemoveAllBlocksWithNoExistingBlock(c *tc.C) {
 
 func (s *stateSuite) TestRemoveAllBlocksWithExistingBlock(c *tc.C) {
 	st := NewState(s.TxnRunnerFactory())
-	err := st.SetBlock(c.Context(), blockcommand.DestroyBlock, "")
+	err := st.SetBlock(c.Context(), int8(blockcommand.DestroyBlock), "")
 	c.Assert(err, tc.ErrorIsNil)
 
 	err = st.RemoveAllBlocks(c.Context())
